Perf: code structure
