+++ ui/src/hooks/useKubernetes.ts
@@ -1,8 +1,3 @@
-import {
-  formatMessage,
-  updateIcon,
-  sendEventToLuna
-} from '@/components/TerminalComponent/helper';
 import { useKubernetesStore } from '@/store/modules/kubernetes.ts';
 import { useTerminalStore } from '@/store/modules/terminal.ts';
 import { base64ToUint8Array, generateWsURL } from './helper';
@@ -14,6 +9,8 @@ import { readText } from 'clipboard-polyfill';
 import { useWebSocket } from '@vueuse/core';
 import { preprocessInput } from '@/utils';
 import { storeToRefs } from 'pinia';
+import { updateIcon } from '@/hooks/helper';
+import { sendEventToLuna, formatMessage } from '@/utils';
 
 import { createDiscreteApi, darkTheme, NIcon } from 'naive-ui';
 import { Cube24Regular } from '@vicons/fluent';
@@ -32,47 +29,34 @@ import xtermTheme from 'xterm-theme';
 import mittBus from '@/utils/mittBus.ts';
 import { MaxTimeout } from '@/config';
 
-const { message, notification } = createDiscreteApi(
-  ['message', 'notification'],
-  {
-    configProviderProps: {
-      theme: darkTheme
-    }
+const { message, notification } = createDiscreteApi(['message', 'notification'], {
+  configProviderProps: {
+    theme: darkTheme
   }
-);
+});
 
 const origin = ref('');
 const lunaId = ref('');
 const counter = ref(0);
 const guaranteeInterval = ref<number | null>(null);
 
-const handleConnected = (
-  socket: WebSocket,
-  pingInterval: Ref<number | null>
-) => {
+const handleConnected = (socket: WebSocket, pingInterval: Ref<number | null>) => {
   const kubernetesStore = useKubernetesStore();
 
   if (pingInterval.value) clearInterval(pingInterval.value);
 
   pingInterval.value = setInterval(() => {
-    if (
-      socket.CLOSED === socket.readyState ||
-      socket.CLOSING === socket.readyState
-    ) {
+    if (socket.CLOSED === socket.readyState || socket.CLOSING === socket.readyState) {
       return clearInterval(pingInterval.value!);
     }
 
     let currentDate: Date = new Date();
 
-    if (
-      kubernetesStore.lastReceiveTime.getTime() - currentDate.getTime() >
-      MaxTimeout
-    ) {
+    if (kubernetesStore.lastReceiveTime.getTime() - currentDate.getTime() > MaxTimeout) {
       message.info('More than 30s do not receive data');
     }
 
-    let pingTimeout: number =
-      currentDate.getTime() - kubernetesStore.lastSendTime.getTime();
+    let pingTimeout: number = currentDate.getTime() - kubernetesStore.lastSendTime.getTime();
 
     if (pingTimeout < 0) return;
 
@@ -86,11 +70,7 @@ const guaranteeLunaConnection = () => {
       guaranteeInterval.value = setInterval(() => {
         counter.value++;
 
-        console.log(
-          '%c DEBUG [ Send Luna PING ]:',
-          'font-size:13px; background: #1ab394; color:#fff;',
-          counter.value
-        );
+        console.log('%c DEBUG [ Send Luna PING ]:', 'font-size:13px; background: #1ab394; color:#fff;', counter.value);
       }, 1000);
     } else {
       clearInterval(guaranteeInterval.value!);
@@ -155,11 +135,7 @@ export const handleInterrupt = (type: string) => {
  * @param label
  * @param isLeaf
  */
-export const setCommonAttributes = (
-  nodes: any,
-  label: string,
-  isLeaf: boolean
-) => {
+export const setCommonAttributes = (nodes: any, label: string, isLeaf: boolean) => {
   const unique = uuid();
 
   Object.assign(nodes, {
@@ -178,12 +154,7 @@ export const setCommonAttributes = (
  * @param namespace
  * @param socket
  */
-export const handleContainer = (
-  containers: any,
-  podName: string,
-  namespace: string,
-  socket: WebSocket
-) => {
+export const handleContainer = (containers: any, podName: string, namespace: string, socket: WebSocket) => {
   const kubernetesStore = useKubernetesStore();
 
   containers.forEach((container: any) => {
@@ -216,8 +187,7 @@ export const handlePods = (pods: any, namespace: string, socket: WebSocket) => {
       pod.isLeaf = false;
       pod.namespace = namespace;
       pod.children = pod.containers;
-      pod.prefix = () =>
-        h(NIcon, { size: 16 }, { default: () => h(Cube24Regular) });
+      pod.prefix = () => h(NIcon, { size: 16 }, { default: () => h(Cube24Regular) });
 
       // 处理最后的 container
       handleContainer(pod.children, pod.name, namespace, socket);
@@ -330,12 +300,7 @@ export const handleTreeMessage = (ws: WebSocket, event: MessageEvent) => {
   }
 };
 
-export const handleTerminalMessage = (
-  ws: WebSocket,
-  event: MessageEvent,
-  createSentry: any,
-  t: any
-) => {
+export const handleTerminalMessage = (ws: WebSocket, event: MessageEvent, createSentry: any, t: any) => {
   const treeStore = useTreeStore();
   const paramsStore = useParamsStore();
   const terminalStore = useTerminalStore();
@@ -385,16 +350,10 @@ export const handleTerminalMessage = (
         }
 
         if (operatedNode.ctrlCAsCtrlZMap) {
-          operatedNode.ctrlCAsCtrlZMap.set(
-            info.k8s_id,
-            sessionInfo.ctrlCAsCtrlZ ? '1' : '0'
-          );
+          operatedNode.ctrlCAsCtrlZMap.set(info.k8s_id, sessionInfo.ctrlCAsCtrlZ ? '1' : '0');
         } else {
           operatedNode.ctrlCAsCtrlZMap = new Map();
-          operatedNode.ctrlCAsCtrlZMap.set(
-            info.k8s_id,
-            sessionInfo.ctrlCAsCtrlZ ? '1' : '0'
-          );
+          operatedNode.ctrlCAsCtrlZMap.set(info.k8s_id, sessionInfo.ctrlCAsCtrlZ ? '1' : '0');
         }
 
         operatedNode.themeName = sessionInfo.themeName;
@@ -426,10 +385,7 @@ export const handleTerminalMessage = (
 
         operatedNode.enableShare = false;
 
-        if (
-          operatedNode.onlineUsersMap &&
-          operatedNode.onlineUsersMap.hasOwnProperty(info.id)
-        ) {
+        if (operatedNode.onlineUsersMap && operatedNode.onlineUsersMap.hasOwnProperty(info.id)) {
           delete operatedNode.onlineUsersMap[info.id];
         }
 
@@ -441,10 +397,7 @@ export const handleTerminalMessage = (
         const data = JSON.parse(info.data);
         const k8s_id: string = info.k8s_id;
 
-        if (
-          operatedNode.onlineUsersMap &&
-          operatedNode.onlineUsersMap[k8s_id]
-        ) {
+        if (operatedNode.onlineUsersMap && operatedNode.onlineUsersMap[k8s_id]) {
           operatedNode.onlineUsersMap[k8s_id].push({
             k8s_id: info.k8s_id,
             ...data
@@ -471,9 +424,7 @@ export const handleTerminalMessage = (
 
         if (operatedNode.onlineUsersMap.hasOwnProperty(k8s_id)) {
           const items = operatedNode?.onlineUsersMap[k8s_id];
-          const index = items.findIndex(
-            (item: any) => item?.terminal_id === data?.terminal_id
-          );
+          const index = items.findIndex((item: any) => item?.terminal_id === data?.terminal_id);
 
           if (index !== -1) {
             items.splice(index, 1);
@@ -501,9 +452,7 @@ export const handleTerminalMessage = (
 
   // 由于 TERMINAL_GET_SHARE_USER 不会返回 k8s id 所以只能根据当前页保存的 k8s id 去获取 node 信息
   if (info.type === 'TERMINAL_GET_SHARE_USER') {
-    const innerOperatedNode = treeStore.getTerminalByK8sId(
-      terminalStore.currentTab
-    );
+    const innerOperatedNode = treeStore.getTerminalByK8sId(terminalStore.currentTab);
     innerOperatedNode.userOptions = JSON.parse(info.data);
 
     treeStore.setK8sIdMap(terminalStore.currentTab, { ...innerOperatedNode });
@@ -565,12 +514,7 @@ export const createConnect = (t: any) => {
  * @param lunaConfig
  * @param socket
  */
-export const initTerminalEvent = (
-  el: HTMLElement,
-  terminal: Terminal,
-  lunaConfig: ILunaConfig,
-  socket: WebSocket
-) => {
+export const initTerminalEvent = (el: HTMLElement, terminal: Terminal, lunaConfig: ILunaConfig, socket: WebSocket) => {
   let fitAddon: FitAddon = new FitAddon();
   let searchAddon: SearchAddon = new SearchAddon();
 
@@ -628,18 +572,11 @@ export const initTerminalEvent = (
   });
 
   terminal.onSelectionChange(() => {
-    terminalStore.setTerminalConfig(
-      'termSelectionText',
-      terminal.getSelection().trim()
-    );
+    terminalStore.setTerminalConfig('termSelectionText', terminal.getSelection().trim());
   });
 
   terminal.attachCustomKeyEventHandler(e => {
-    if (
-      e.altKey &&
-      e.shiftKey &&
-      (e.key === 'ArrowRight' || e.key === 'ArrowLeft')
-    ) {
+    if (e.altKey && e.shiftKey && (e.key === 'ArrowRight' || e.key === 'ArrowLeft')) {
       switch (e.key) {
         case 'ArrowRight':
           mittBus.emit('alt-shift-right');
@@ -697,8 +634,7 @@ export const initElEvent = (
       try {
         text = await readText();
       } catch (e) {
-        if (terminalStore.termSelectionText !== '')
-          text = terminalStore.termSelectionText;
+        if (terminalStore.termSelectionText !== '') text = terminalStore.termSelectionText;
       } finally {
         socket.send(
           JSON.stringify({
@@ -764,10 +700,7 @@ export const sendK8sMessage = (socket: WebSocket, type: string, data: any) => {
   );
 };
 
-export const initMittBusEvents = (
-  searchAddon: SearchAddon,
-  socket: WebSocket
-) => {
+export const initMittBusEvents = (searchAddon: SearchAddon, socket: WebSocket) => {
   mittBus.on('terminal-search', ({ keyword, type = '' }) => {
     const searchOption: ISearchOptions = {
       caseSensitive: false,
@@ -812,11 +745,7 @@ export const initMittBusEvents = (
 /**
  * @description 创建 K8s 终端
  */
-export const createTerminal = (
-  el: HTMLElement,
-  socket: WebSocket,
-  lunaConfig: ILunaConfig
-) => {
+export const createTerminal = (el: HTMLElement, socket: WebSocket, lunaConfig: ILunaConfig) => {
   const { fontSize, lineHeight, fontFamily } = lunaConfig;
 
   const options = {
@@ -833,12 +762,7 @@ export const createTerminal = (
 
   const terminal: Terminal = new Terminal(options);
 
-  const { fitAddon, searchAddon } = initTerminalEvent(
-    el,
-    terminal,
-    lunaConfig,
-    socket
-  );
+  const { fitAddon, searchAddon } = initTerminalEvent(el, terminal, lunaConfig, socket);
 
   initElEvent(el, terminal, fitAddon, socket, lunaConfig);
   initCustomWindowEvent(fitAddon);
