+++ ui/src/hooks/useTerminal.ts
@@ -33,13 +33,9 @@ import {
   onWebsocketOpen,
   onWebsocketWrong
 } from '@/hooks/helper';
-import {
-  formatMessage,
-  sendEventToLuna,
-  updateIcon,
-  wsIsActivated
-} from '@/components/TerminalComponent/helper';
 import mittBus from '@/utils/mittBus.ts';
+import { updateIcon } from '@/hooks/helper';
+import { sendEventToLuna, formatMessage, wsIsActivated } from '@/utils';
 
 enum MessageType {
   PING = 'PING',
@@ -65,12 +61,7 @@ interface ICallbackOptions {
   transSocket?: WebSocket;
 
   // emit 事件
-  emitCallback?: (
-    e: string,
-    type: string,
-    msg: any,
-    terminal?: Terminal
-  ) => void;
+  emitCallback?: (e: string, type: string, msg: any, terminal?: Terminal) => void;
 
   // t
   i18nCallBack?: (key: string) => string;
@@ -78,10 +69,7 @@ interface ICallbackOptions {
 
 const { message } = createDiscreteApi(['message']);
 
-export const useTerminal = async (
-  el: HTMLElement,
-  option: ICallbackOptions
-): Promise<ITerminalInstance> => {
+export const useTerminal = async (el: HTMLElement, option: ICallbackOptions): Promise<ITerminalInstance> => {
   let sentry: Sentry;
   let socket: WebSocket;
   let terminal: Terminal | undefined;
@@ -148,13 +136,7 @@ export const useTerminal = async (
 
         updateIcon(info.setting);
 
-        socket.send(
-          formatMessage(
-            terminalId.value,
-            'TERMINAL_INIT',
-            JSON.stringify(terminalData)
-          )
-        );
+        socket.send(formatMessage(terminalId.value, 'TERMINAL_INIT', JSON.stringify(terminalData)));
         break;
       }
       case MessageType.CLOSE: {
@@ -196,8 +178,7 @@ export const useTerminal = async (
         break;
       }
       case MessageType.TERMINAL_SHARE_USER_REMOVE: {
-        option.i18nCallBack &&
-          message.info(option.i18nCallBack('RemoveShareUser'));
+        option.i18nCallBack && message.info(option.i18nCallBack('RemoveShareUser'));
         socket.close();
         break;
       }
@@ -205,8 +186,7 @@ export const useTerminal = async (
       }
     }
 
-    option.emitCallback &&
-      option.emitCallback('socketData', msg.type, msg, terminal);
+    option.emitCallback && option.emitCallback('socketData', msg.type, msg, terminal);
   };
 
   /**
@@ -232,11 +212,7 @@ export const useTerminal = async (
   /**
    * 设置主题
    */
-  const setTerminalTheme = (
-    themeName: string,
-    terminal: Terminal,
-    emits: any
-  ) => {
+  const setTerminalTheme = (themeName: string, terminal: Terminal, emits: any) => {
     const theme = xtermTheme[themeName] || defaultTheme;
 
     terminal.options.theme = theme;
@@ -274,12 +250,7 @@ export const useTerminal = async (
           }
         }
       } else {
-        writeBufferToTerminal(
-          enableZmodem.value,
-          zmodemStatus.value,
-          terminal!,
-          event.data
-        );
+        writeBufferToTerminal(enableZmodem.value, zmodemStatus.value, terminal!, event.data);
       }
     } else {
       dispatch(event.data);
@@ -370,13 +341,7 @@ export const useTerminal = async (
     el.addEventListener(
       'contextmenu',
       (e: MouseEvent) => {
-        handleContextMenu(
-          e,
-          lunaConfig,
-          socket!,
-          terminalId.value,
-          termSelectionText.value
-        );
+        handleContextMenu(e, lunaConfig, socket!, terminalId.value, termSelectionText.value);
       },
       false
     );
@@ -531,11 +496,7 @@ export const useTerminal = async (
     return terminalInstance;
   };
 
-  const initializeTerminal = (
-    terminal: Terminal,
-    socket: WebSocket,
-    type: string
-  ) => {
+  const initializeTerminal = (terminal: Terminal, socket: WebSocket, type: string) => {
     initElEvent();
     initTerminalEvent();
     initCustomWindowEvent();
@@ -600,15 +561,9 @@ export const useTerminal = async (
 
     lunaConfig = terminalStore.getConfig;
 
-    const [socketResult, terminalResult] = await Promise.allSettled([
-      createSocket(),
-      createTerminal(lunaConfig)
-    ]);
+    const [socketResult, terminalResult] = await Promise.allSettled([createSocket(), createTerminal(lunaConfig)]);
 
-    if (
-      socketResult.status === 'fulfilled' &&
-      terminalResult.status === 'fulfilled'
-    ) {
+    if (socketResult.status === 'fulfilled' && terminalResult.status === 'fulfilled') {
       socket = socketResult.value!;
       terminal = terminalResult.value;
